+++ src/common/Path.ts
@@ -1,20 +1,20 @@
-import * as PATH from 'path'
-import * as vscode from 'vscode'
+import * as PATH from 'path';
+import * as vscode from 'vscode';
 
 const { file } = vscode.Uri,
-    { getWorkspaceFolder, asRelativePath } = vscode.workspace
+    { getWorkspaceFolder, asRelativePath } = vscode.workspace;
 
 export default class Path {
 
-    private _fsPath: string
-    private _root: string
-    private _relative: string
-    private _partitions: string[]
+    private _fsPath: string;
+    private _root: string;
+    private _relative: string;
+    private _partitions: string[];
 
     constructor(fsPath: string) {
         if (process.platform === 'win32')
-            fsPath = fsPath.replace(/\\/g, '\\\\')
-        this._fsPath = fsPath
+            fsPath = fsPath.replace(/\\/g, '\\\\');
+        this._fsPath = fsPath;
     }
     /**
      * Get file's absolute path.
@@ -22,7 +22,7 @@ export default class Path {
      * @memberof Path
      */
     fsPath(): string {
-        return this._fsPath
+        return this._fsPath;
     }
     /**
      * Get file's workspace in vscode.
@@ -31,8 +31,8 @@ export default class Path {
      */
     root(): string {
         if (!this._root)
-            this._root = getWorkspaceFolder(file(this._fsPath)).uri.fsPath
-        return this._root
+            this._root = getWorkspaceFolder(file(this._fsPath)).uri.fsPath;
+        return this._root;
     }
     /**
      * Get file's relative path from workspace in vscode.
@@ -41,8 +41,8 @@ export default class Path {
      */
     asRelative(): string {
         if (!this._relative)
-            this._relative = asRelativePath(this._fsPath, false)
-        return this._relative
+            this._relative = asRelativePath(this._fsPath, false);
+        return this._relative;
     }
     /**
      * Get different partitions of file's relative path.
@@ -51,10 +51,10 @@ export default class Path {
      */
     partitions(): string[] {
         if (!this._partitions) {
-            let rPathPartitions = this.asRelative().match(/(.*?)[\/\\]?(([^\/\\]+?)\.(\w+))$/)
-            this._partitions = [this.fsPath(), this.root(), ...rPathPartitions]
+            let rPathPartitions = this.asRelative().match(/(.*?)[\/\\]?(([^\/\\]+?)\.(\w+))$/);
+            this._partitions = [this.fsPath(), this.root(), ...rPathPartitions];
         }
-        return this._partitions
+        return this._partitions;
     }
     /**
      * Wrap the file or directory name in path with double quote, if the name include whitespace.
@@ -63,16 +63,16 @@ export default class Path {
     static wrapWhiteSpace(path: string): string {
         return path.replace(/(^|[\/\\])([^\/\\]+)(?=[\/\\]|$)/g, (match, $0, $1) => {
             if (!/"/.test($1) && /\s/.test($1))
-                return `${$0}"${$1}"`
+                return `${$0}"${$1}"`;
             else
-                return match
-        })
+                return match;
+        });
     }
     /**
      * Unified separator in path as '\' in win32 or '/' in posix.
      * @static
      */
     static unifiedSeparator(path: string): string {
-        return path.replace(/[\/\\]/g, PATH.sep)
+        return path.replace(/[\/\\]/g, PATH.sep);
     }
 }