+++ src/util/formatter.ts
@@ -1,5 +1,5 @@
 export function format(template: string, dataObject): string {
     return template.replace(/\$\{([^}]+)\}/g, (match, variable) => {
-        return dataObject[variable]
-    })
+        return dataObject[variable];
+    });
 }