+++ test/files-to-run/run as.spec.ts
@@ -1,5 +1,0 @@
-describe('Run As', () => {
-    it('Right click *.spec.ts but actually run *.spec.js in Mocha', () => {
-        // Pass
-    })
-})